requirements.txt
black
isort
mypy
slack_bolt==1.18.0
python-dotenv==1.0.0
langchain==0.0.292
openai==0.28.0
momento==1.9.2